% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFperformance.R
\name{GOFperformance}
\alias{GOFperformance}
\title{Output Performance comparisons for simulated and observed hydrological time series data}
\usage{
GOFperformance(sim, obs)
}
\arguments{
\item{sim}{numeric 'data.frame', 'matrix', or 'vector' with simulated values}

\item{obs}{numeric 'data.frame', 'matrix', or 'vector' with observed values}
}
\description{
Performes Nash-Sutcliffe Efficiency, relative Nash-Sutcliffe Efficiency, Root Mean Square Error, Relative Root Mean Square Error, Ratio of RMSE to the standard deviation of the observations, Percent Bias, and Kling-Gupta Efficiency
}
