% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr4jSCE.R
\name{gr4jSCE}
\alias{gr4jSCE}
\title{Runs SCE Calibration}
\usage{
gr4jSCE(inputdata, seed, x1 = log(c(1, 5000)), x2 = asinh(c(-10, 5)),
  x3 = log(c(1, 500)), x4 = log(c(0.5, 4) - 0.5), transformed = FALSE,
  warmup = 365, ncomplex = 20)
}
\arguments{
\item{inputdata}{Input data for GR4J model (requires Data$Q, Data$P, and Data$E)}

\item{seed}{Random number for sampling within SCE}

\item{x1}{X1 range}

\item{x2}{X2 range}

\item{x3}{X3 range}

\item{x4}{X4 range}

\item{transformed}{Transformed parameters before use to improve identifiability}

\item{warmup}{Number of Days to Warmup for GR4J}

\item{ncomplex}{Number of Complexes for SCE}
}
\description{
Runs SCE Calibration
}
