% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr4jSim.R
\name{gr4jSim}
\alias{gr4jSim}
\title{Outputs GR4J modelled run with observed flow}
\usage{
gr4jSim(input, x1, x2, x3, x4, area, etmult = 1, transformed = TRUE,
  warmup = 365)
}
\arguments{
\item{input}{Input data (requires Data$Q, Data$P, and Data$E) (mm/day)}

\item{x1}{Value for X1 Parameter in GR4J}

\item{x2}{Value for X2 Parameter in GR4J}

\item{x3}{Value for X3 Parameter in GR4J}

\item{x4}{Value for X4 Parameter in GR4J}

\item{etmult}{Multiplier applied to Data$E to estimate potential evapotranspiration (default is 1)}

\item{warmup}{Number of Days to Warmup the hydrologic model}

\item{trasnformed}{Transformed parameters before use to improve identifiability}
}
\description{
Outputs GR4J modelled run with observed flow
}
