% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihacrescmdexpuhSim.R
\name{ihacrescmdexpuhSim}
\alias{ihacrescmdexpuhSim}
\title{Outputs IHACRES CMD modelled run with observed flow
 
 Runs IHACRES CMD model with expuh routing model}
\usage{
ihacrescmdexpuhSim(input, delay = 0, f, e = 1, d, shape, tau_s, tau_q, v_s,
  warmup = 365)
}
\arguments{
\item{input}{Input data (requires Data$P, and Data$E) (mm/day)}

\item{delay}{Time-delay for hydrologic routing}

\item{f}{CMD stress threshold as a proportion of d}

\item{e}{Temperature to PET conversion factor}

\item{d}{CMD threshold producing flow}

\item{shape}{shape}

\item{tau_s}{time constraints}

\item{tau_q}{time constraints}

\item{v_s}{fractional volume}

\item{warmup}{Number of Days to Warmup the hydrologic model}
}
\description{
Outputs IHACRES CMD modelled run with observed flow
 
 Runs IHACRES CMD model with expuh routing model
}
