% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrmse.R
\name{rrmse}
\alias{rrmse}
\title{Outputs relative root mean square error (modified from hydroGOF's RMSE function)}
\usage{
rrmse(sim, obs, na.rm = TRUE, ...)
}
\arguments{
\item{sim}{numeric 'data.frame', 'matrix' or 'vector' with simulated values}

\item{obs}{numeric 'data.frame', 'matrix' or 'vector' with simulated values}

\item{na.rm}{a logical value indicating whether 'NA' should be stripped before the computation proceeds.}
}
\description{
Outputs relative root mean square error (modified from hydroGOF's RMSE function)
}
