% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacramentoexpuhSim.R
\name{sacramentoexpuhSim}
\alias{sacramentoexpuhSim}
\title{Outputs Sacramento modelled run with observed flow}
\usage{
sacramentoexpuhSim(input, delay = 0, uztwm, uzfwm, uzk, pctim, adimp, zperc,
  rexp, lztwm, lzfsm, lzfpm, lzsk, lzpk, pfree, etmult = 1, tau_s, tau_q, v_s,
  warmup = 365)
}
\arguments{
\item{input}{Input data (requires Data$Q, Data$P, and Data$E) (mm/day)}

\item{delay}{Time-delay for hydrologic routing}

\item{uztwm}{Upper zone maximum tension water capacity (mm)}

\item{uzfwm}{Upper zone maximum free water capacity (mm).}

\item{uzk}{Lateral drainage rate of upper zone free water expressed as a fraction of contents per day.}

\item{pctim}{The fraction of the catchment which produces impervious runoff during low flow conditions.}

\item{adimp}{The additional fraction of the catchment which exhibits impervious characteristics when the catchment's tension water requirements are met.}

\item{zperc}{Maximum percolation (from upper zone free water into the lower zone) rate coefficient.}

\item{rexp}{An exponent determining the rate of change of the percolation rate with changing lower zone water contents.}

\item{lztwm}{Lower zone maximum tension water capacity (mm)}

\item{lzfsm}{Lower zone maximum supplemental free water capacity (mm)}

\item{lzfpm}{Lower zone maximum primary free water capacity (mm)}

\item{lzsk}{Lateral drainage rate of lower zone supplemental free water expressed as a fraction of contents per day}

\item{lzpk}{Lateral drainage rate of lower zone primary free water expressed as a fraction of contents per day}

\item{pfree}{Direct percolation fraction from upper to lower zone free water (the percentage of percolated water which is available to the lower zone free water aquifers before all lower zone tension water deficiencies are satisfied)}

\item{etmult}{Multiplier applied to Data$E to estimate potential evapotranspiration (default is 1)}

\item{tau_s}{time constraints}

\item{tau_q}{time constraints}

\item{v_s}{fractional volume}

\item{warmup}{Number of Days to Warmup the hydrologic model}
}
\description{
Runs Sacramento model with expuh routing model
}
