% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snucahl.R
\docType{package}
\name{snucahl-package}
\alias{snucahl}
\alias{snucahl-package}
\title{SNU-CAHL Package}
\description{
Maintained by Seoul Nationaln University Research Group for Climate Change Adaptation in Water Resources.  Visit <http://ccawr.snu.ac.kr/> for our homepage.
}
\details{
Functions are automatically imported from the library. Useful scripts are located in `scripts/` folder and examples are present in the `examples/` folder.

The following functions are available
\code{\link{GOFperformance}} 
\code{\link{gr4jSCE}}
\code{\link{gr4jSim}}
\code{\link{ihacrescmdexpuhSCE}}
\code{\link{ihacrescmdexpuhSim}}
\code{\link{rrmse}}
\code{\link{sacramentoexpuhSCE}}
\code{\link{sacramentoexpuhSim}}
}
