% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr4jSCE.R
\name{gr4jSCE}
\alias{gr4jSCE}
\title{Runs SCE Calibration for GR4J Model}
\usage{
gr4jSCE(inputdata, seed, x1 = c(1, 5000), x2 = c(-10, 10), x3 = c(1,
  1000), x4 = c(0.5, 4), warmup = 365, ncomplex = 20, maxit = 200,
  transformed = TRUE)
}
\arguments{
\item{inputdata}{Input data for GR4J model (requires Data$Q, Data$P, and Data$E) (mm/day)}

\item{seed}{Random number for sampling within SCE}

\item{x1}{X1 range}

\item{x2}{X2 range}

\item{x3}{X3 range}

\item{x4}{X4 range}

\item{warmup}{Number of Days to Warmup for GR4J}

\item{ncomplex}{Number of Complexes for SCE}

\item{maxit}{Maximum number of iterations}

\item{transformed}{GR4J Values already transformed for model performance}
}
\description{
Outputs GR4J Parameters (non-transformed) and NSE value.
}
