% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr4jSCE.R
\name{gr4jSCE}
\alias{gr4jSCE}
\title{Runs SCE Calibration for GR4J Model}
\usage{
gr4jSCE(inputdata, seed, x1 = c(100, 1200), x2 = c(-5, 3), x3 = c(20,
  300), x4 = c(1.1, 3), warmup = 365, ncomplex = 20, maxit = 200,
  transformed = FALSE)
}
\arguments{
\item{inputdata}{Input data (requires Data$Q, Data$P, and Data$E) (mm/day)}

\item{seed}{Random number for sampling within SCE}

\item{x1}{X1 range}

\item{x2}{X2 range}

\item{x3}{X3 range}

\item{x4}{X4 range}

\item{warmup}{Number of Days to Warmup the hydrologic model}

\item{ncomplex}{Number of Complexes for SCE}

\item{maxit}{Maximum number of iterations}

\item{transformed}{GR4J Values already transformed for model performance}
}
\description{
Outputs GR4J Parameters (non-transformed) and NSE value.
}
